do org upload token separately
