+++ src/services/orgUploadToken/useRegenerateOrgUploadToken.test.jsx
@@ -2,14 +2,15 @@ import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
 import { renderHook, waitFor } from '@testing-library/react'
 import { graphql, HttpResponse } from 'msw'
 import { setupServer } from 'msw/node'
-import React from 'react'
 import { MemoryRouter, Route } from 'react-router-dom'
 
 import { useRegenerateOrgUploadToken } from './useRegenerateOrgUploadToken'
 
 const data = {
-  regenerateOrgUploadToken: {
-    orgUploadToken: 'new token',
+  data: {
+    regenerateOrgUploadToken: {
+      orgUploadToken: 'new token',
+    },
   },
 }
 
@@ -22,7 +23,7 @@ afterAll(() => server.close())
 const queryClient = new QueryClient({
   defaultOptions: { queries: { retry: false } },
 })
-const wrapper: React.FC<React.PropsWithChildren> = ({ children }) => (
+const wrapper = ({ children }) => (
   <MemoryRouter initialEntries={['/gh/codecov/gazebo']}>
     <Route path="/:provider/:owner/:repo">
       <QueryClientProvider client={queryClient}>{children}</QueryClientProvider>
